# api/openapi/service.yaml
# Generated by `make openapi-generate` from the Rust contract; do not edit by hand.
openapi: 3.1.0
info:
  title: rust-service-template-rest
  description: 'HTTP contract served by the `service` binary: platform probes and feature operations, with RFC 9457 problem responses from one closed catalog.'
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
servers:
- url: /
  description: current service origin
paths:
  /health/live:
    get:
      tags:
      - system
      summary: Liveness probe
      operationId: healthLive
      responses:
        '200':
          description: ok
          content:
            text/plain:
              schema:
                type: string
              example: ok
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/RequestEntityTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security: []
      x-security-decision:
        exposure: public
        rationale: process-only platform liveness endpoint with no dependency details
  /health/ready:
    get:
      tags:
      - system
      summary: Readiness probe
      operationId: healthReady
      responses:
        '200':
          description: ready
          content:
            text/plain:
              schema:
                type: string
              example: ok
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/RequestEntityTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          description: not ready
          content:
            text/plain:
              schema:
                type: string
              example: not ready
      security: []
      x-security-decision:
        exposure: public
        rationale: platform readiness endpoint exposing only generic ready state
  /webhooks/{endpoint_id}:
    post:
      tags:
      - webhooks
      summary: Accept a signed Standard Webhooks delivery
      description: The required Standard Webhooks signature headers authenticate this public endpoint. A successful response acknowledges durable asynchronous ownership only.
      operationId: receiveWebhook
      parameters:
      - name: endpoint_id
        in: path
        description: operator-configured receiving endpoint ID
        required: true
        schema:
          type: string
      - name: webhook-id
        in: header
        description: required Standard Webhooks message ID, 1–255 bytes without a dot
        required: true
        schema:
          type: string
      - name: webhook-timestamp
        in: header
        description: required Standard Webhooks timestamp
        required: true
        schema:
          type: string
      - name: webhook-signature
        in: header
        description: required Standard Webhooks signature candidates
        required: true
        schema:
          type: string
      requestBody:
        description: Original signed binary body, at most 128 KiB. No JSON decoding or payload schema is required.
        content:
          '*/*': {}
      responses:
        '204':
          description: webhook receipt is durable
        '400':
          $ref: '#/components/responses/WebhookRejected'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/RequestEntityTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
      security: []
components:
  schemas:
    InvalidParam:
      type: object
      description: |-
        Which part of a request failed validation, following the RFC 9457
        extension-member example. Never carries the submitted value.
      required:
      - name
      - reason
      properties:
        name:
          type: string
          description: |-
            An RFC 6901 JSON pointer for a request-body member, or a
            `location.name` pair such as `query.limit` for a parameter.
          example: /slug
        reason:
          type: string
          description: |-
            Why the contract rejected it. Names the constraint that failed and
            never echoes the submitted value.
          example: maximum string length is 64
      additionalProperties: false
    Problem:
      type: object
      description: |-
        RFC 9457 problem details: the failure envelope of every non-success
        response, keyed by a stable `code` from one closed catalog.
      required:
      - code
      - type
      - title
      - status
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          example: bad_request
        detail:
          type: string
          example: invalid request framing
        instance:
          type: string
          format: uri-reference
          description: |-
            URI reference identifying this occurrence when the service exposes
            one.
        invalid_params:
          type: array
          items:
            $ref: '#/components/schemas/InvalidParam'
          description: |-
            Which parts of the request failed validation, following the RFC 9457
            extension-member example. Present only on a validation rejection,
            and never carrying the submitted value.
        request_id:
          type: string
          example: 9ccecdfd-92c9-4665-a464-06f8ba73cd77
        status:
          type: integer
          format: int32
          example: 400
          minimum: 0
        title:
          type: string
          example: bad request
        type:
          type: string
          format: uri-reference
          description: Stable URI reference identifying the problem class.
          example: https://www.rfc-editor.org/rfc/rfc9110#section-15.5.1
      additionalProperties: false
  responses:
    AuthenticationForbidden:
      description: downstream authorization denied the verified principal
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationMalformed:
      description: bearer authentication is malformed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationOversize:
      description: bearer authentication exceeds the admitted header value size
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationTimeout:
      description: the enclosing request budget expired before a response committed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationUnauthorized:
      description: bearer authentication is required or invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationUnavailable:
      description: bearer authentication trust or provider is unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    BadRequest:
      description: request is malformed or invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    IdempotencyBadRequest:
      description: |-
        bearer authentication is malformed, the Idempotency-Key header is missing
        or invalid, or the request is otherwise malformed or invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    IdempotencyKeyMismatch:
      description: |-
        the Idempotency-Key is bound to a different request, or the operation
        cannot process the request content
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    IdempotencyRequestInProgress:
      description: |-
        a request with this Idempotency-Key is still in progress, or the operation
        reports a conflict; retry with the same key after the Retry-After interval
      headers:
        Retry-After:
          schema:
            type: integer
            format: int64
            minimum: 0
          description: seconds to wait before retrying with the same Idempotency-Key
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    IdempotencyUnavailable:
      description: |-
        bearer authentication trust or idempotent request processing is unavailable,
        or the committed outcome is uncertain; retry with the same Idempotency-Key
      headers:
        Retry-After:
          schema:
            type: integer
            format: int64
            minimum: 0
          description: seconds to wait before retrying with the same Idempotency-Key
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: unexpected server failure
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: webhook endpoint is unknown
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RequestEntityTooLarge:
      description: request body exceeds configured limit
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: durable webhook receipt storage is unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    WebhookRejected:
      description: webhook body, message identity, or signature evidence is unacceptable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
- bearerAuth: []
tags:
- name: system
  description: Operational endpoints for liveness and readiness.
- name: webhooks
  description: Authenticated inbound webhook admission.